move utility functions for matrix plotting to nilearn.matrix._utils
