+++ nilearn/plotting/matrix/_utils.py
@@ -1,0 +1,35 @@
+import numpy as np
+from scipy.cluster.hierarchy import leaves_list, linkage, optimal_leaf_ordering
+
+
+def _mask_matrix(mat, tri):
+    """Help for plot_matrix.
+
+    This function masks the matrix depending on the provided
+    value of ``tri``.
+    """
+    if tri == "lower":
+        mask = np.tri(mat.shape[0], k=-1, dtype=bool) ^ True
+    else:
+        mask = np.tri(mat.shape[0], dtype=bool) ^ True
+    return np.ma.masked_array(mat, mask)
+
+
+def _reorder_matrix(mat, labels, reorder):
+    """Help for plot_matrix.
+
+    This function reorders the provided matrix.
+    """
+    if not labels:
+        raise ValueError("Labels are needed to show the reordering.")
+
+    linkage_matrix = linkage(mat, method=reorder)
+    ordered_linkage = optimal_leaf_ordering(linkage_matrix, mat)
+    index = leaves_list(ordered_linkage)
+    # make sure labels is an ndarray and copy it
+    labels = np.array(labels).copy()
+    mat = mat.copy()
+    # and reorder labels and matrix
+    labels = labels[index].tolist()
+    mat = mat[index, :][:, index]
+    return mat, labels
